+++ src/main/java/me/lambdaurora/lambdynlights/LambDynLights.java
@@ -31,6 +31,8 @@ import net.minecraft.resource.SynchronousResourceReloader;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.MathHelper;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.RenderWorldLastEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
@@ -54,7 +56,7 @@ import java.util.function.Predicate;
  */
 @Mod(LambDynLights.MODID)
 public class LambDynLights {
-    public static final String MODID = "lambdynlights";
+    public static final String MODID = "ryoamiclights";
     private static final double MAX_RADIUS = 7.75;
     private static LambDynLights INSTANCE;
     public final Logger logger = LogManager.getLogger(MODID);
@@ -66,6 +68,7 @@ public class LambDynLights {
     public LambDynLights() {
         ModLoadingContext.get().registerExtensionPoint(ExtensionPoint.DISPLAYTEST, () -> Pair.of(() -> FMLNetworkConstants.IGNORESERVERONLY, (a, b) -> true));
         EnvExecutor.runInEnv(Dist.CLIENT, () -> this::onInitializeClient);
+        EnvExecutor.runInEnv(Dist.CLIENT, () -> () -> Mod.EventBusSubscriber.Bus.FORGE.bus().get().addListener(this::renderWorldLastEvent));
     }
 
     public void onInitializeClient() {
@@ -78,6 +81,12 @@ public class LambDynLights {
         Platform.getMod(MODID).registerConfigurationScreen(SettingsScreen::new);
 
         DynamicLightHandlers.registerDefaultHandlers();
+    }
+
+    @SubscribeEvent
+    public void renderWorldLastEvent(RenderWorldLastEvent event) {
+        MinecraftClient.getInstance().getProfiler().push("dynamic_lighting");
+        get().updateAll(event.getContext());
     }
 
     /**
@@ -95,7 +104,7 @@ public class LambDynLights {
             this.lastUpdateCount = 0;
 
             for (DynamicLightSource lightSource : this.dynamicLightSources) {
-                if (lightSource.lambdynlights_updateDynamicLight(renderer)) this.lastUpdateCount++;
+                if (lightSource.ryoamicLights$updateDynamicLight(renderer)) this.lastUpdateCount++;
             }
         }
     }
@@ -129,7 +138,7 @@ public class LambDynLights {
      */
     public int getLightmapWithDynamicLight(@NotNull Entity entity, int lightmap) {
         int posLightLevel = (int) this.getDynamicLightLevel(entity.getBlockPos());
-        int entityLuminance = ((DynamicLightSource) entity).getLuminance();
+        int entityLuminance = ((DynamicLightSource) entity).ryoamicLights$getLuminance();
 
         return this.getLightmapWithDynamicLight(Math.max(posLightLevel, entityLuminance), lightmap);
     }
@@ -182,12 +191,12 @@ public class LambDynLights {
      * @return the dynamic light level at the specified position
      */
     public static double maxDynamicLightLevel(@NotNull BlockPos pos, @NotNull DynamicLightSource lightSource, double currentLightLevel) {
-        int luminance = lightSource.getLuminance();
+        int luminance = lightSource.ryoamicLights$getLuminance();
         if (luminance > 0) {
             // Can't use Entity#squaredDistanceTo because of eye Y coordinate.
-            double dx = pos.getX() - lightSource.getDynamicLightX() + 0.5;
-            double dy = pos.getY() - lightSource.getDynamicLightY() + 0.5;
-            double dz = pos.getZ() - lightSource.getDynamicLightZ() + 0.5;
+            double dx = pos.getX() - lightSource.ryoamicLights$getDynamicLightX() + 0.5;
+            double dy = pos.getY() - lightSource.ryoamicLights$getDynamicLightY() + 0.5;
+            double dz = pos.getZ() - lightSource.ryoamicLights$getDynamicLightZ() + 0.5;
 
             double distance = Math.sqrt(dx * dx + dy * dy + dz * dz);
             // 7.75 because else we would have to update more chunks and that's not a good idea.
@@ -209,7 +218,7 @@ public class LambDynLights {
      * @param lightSource the light source to add
      */
     public void addLightSource(@NotNull DynamicLightSource lightSource) {
-        if (!lightSource.getDynamicLightWorld().isClient())
+        if (!lightSource.ryoamicLights$getDynamicLightWorld().isClient())
             return;
         if (!this.config.getDynamicLightsMode().isEnabled())
             return;
@@ -225,7 +234,7 @@ public class LambDynLights {
      * @return {@code true} if the light source is tracked, else {@code false}
      */
     public boolean containsLightSource(@NotNull DynamicLightSource lightSource) {
-        if (!lightSource.getDynamicLightWorld().isClient())
+        if (!lightSource.ryoamicLights$getDynamicLightWorld().isClient())
             return false;
         return this.dynamicLightSources.contains(lightSource);
     }
@@ -252,7 +261,7 @@ public class LambDynLights {
             if (it.equals(lightSource)) {
                 dynamicLightSources.remove();
                 if (MinecraftClient.getInstance().worldRenderer != null)
-                    lightSource.lambdynlights_scheduleTrackedChunksRebuild(MinecraftClient.getInstance().worldRenderer);
+                    lightSource.ryoamicLights$scheduleTrackedChunksRebuild(MinecraftClient.getInstance().worldRenderer);
                 break;
             }
         }
@@ -268,9 +277,9 @@ public class LambDynLights {
             it = dynamicLightSources.next();
             dynamicLightSources.remove();
             if (MinecraftClient.getInstance().worldRenderer != null) {
-                if (it.getLuminance() > 0)
-                    it.resetDynamicLight();
-                it.lambdynlights_scheduleTrackedChunksRebuild(MinecraftClient.getInstance().worldRenderer);
+                if (it.ryoamicLights$getLuminance() > 0)
+                    it.ryoamicLights$resetDynamicLight();
+                it.ryoamicLights$scheduleTrackedChunksRebuild(MinecraftClient.getInstance().worldRenderer);
             }
         }
     }
@@ -288,9 +297,9 @@ public class LambDynLights {
             if (filter.test(it)) {
                 dynamicLightSources.remove();
                 if (MinecraftClient.getInstance().worldRenderer != null) {
-                    if (it.getLuminance() > 0)
-                        it.resetDynamicLight();
-                    it.lambdynlights_scheduleTrackedChunksRebuild(MinecraftClient.getInstance().worldRenderer);
+                    if (it.ryoamicLights$getLuminance() > 0)
+                        it.ryoamicLights$resetDynamicLight();
+                    it.ryoamicLights$scheduleTrackedChunksRebuild(MinecraftClient.getInstance().worldRenderer);
                 }
                 break;
             }
@@ -391,9 +400,9 @@ public class LambDynLights {
      * @param lightSource the light source
      */
     public static void updateTracking(@NotNull DynamicLightSource lightSource) {
-        if (!lightSource.isDynamicLightEnabled() && lightSource.getLuminance() > 0) {
+        if (!lightSource.isDynamicLightEnabled() && lightSource.ryoamicLights$getLuminance() > 0) {
             lightSource.setDynamicLightEnabled(true);
-        } else if (lightSource.isDynamicLightEnabled() && lightSource.getLuminance() < 1) {
+        } else if (lightSource.isDynamicLightEnabled() && lightSource.ryoamicLights$getLuminance() < 1) {
             lightSource.setDynamicLightEnabled(false);
         }
     }