+++ src/main/java/me/lambdaurora/lambdynlights/mixin/lightsource/BlockEntityMixin.java
@@ -18,7 +18,6 @@ import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.WorldRenderer;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.ChunkSectionPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.World;
@@ -49,25 +48,25 @@ public abstract class BlockEntityMixin implements DynamicLightSource
     private LongOpenHashSet trackedLitChunkPos = new LongOpenHashSet();
 
     @Override
-    public double getDynamicLightX()
+    public double ryoamicLights$getDynamicLightX()
     {
         return this.pos.getX() + 0.5;
     }
 
     @Override
-    public double getDynamicLightY()
+    public double ryoamicLights$getDynamicLightY()
     {
         return this.pos.getY() + 0.5;
     }
 
     @Override
-    public double getDynamicLightZ()
+    public double ryoamicLights$getDynamicLightZ()
     {
         return this.pos.getZ() + 0.5;
     }
 
     @Override
-    public World getDynamicLightWorld()
+    public World ryoamicLights$getDynamicLightWorld()
     {
         return this.world;
     }
@@ -79,13 +78,13 @@ public abstract class BlockEntityMixin implements DynamicLightSource
     }
 
     @Override
-    public void resetDynamicLight()
+    public void ryoamicLights$resetDynamicLight()
     {
         this.lambdynlights_lastLuminance = 0;
     }
 
     @Override
-    public void dynamicLightTick()
+    public void ryoamicLights$dynamicLightTick()
     {
         // We do not want to update the entity on the server.
         if (this.world == null || !this.world.isClient())
@@ -101,13 +100,13 @@ public abstract class BlockEntityMixin implements DynamicLightSource
     }
 
     @Override
-    public int getLuminance()
+    public int ryoamicLights$getLuminance()
     {
         return this.lambdynlights_luminance;
     }
 
     @Override
-    public boolean shouldUpdateDynamicLight()
+    public boolean ryoamicLights$shouldUpdateDynamicLight()
     {
         DynamicLightsMode mode = LambDynLights.get().config.getDynamicLightsMode();
         if (!mode.isEnabled())
@@ -124,12 +123,12 @@ public abstract class BlockEntityMixin implements DynamicLightSource
     }
 
     @Override
-    public boolean lambdynlights_updateDynamicLight(@NotNull WorldRenderer renderer)
+    public boolean ryoamicLights$updateDynamicLight(@NotNull WorldRenderer renderer)
     {
-        if (!this.shouldUpdateDynamicLight())
+        if (!this.ryoamicLights$shouldUpdateDynamicLight())
             return false;
 
-        int luminance = this.getLuminance();
+        int luminance = this.ryoamicLights$getLuminance();
 
         if (luminance != this.lambdynlights_lastLuminance) {
             this.lambdynlights_lastLuminance = luminance;
@@ -161,14 +160,14 @@ public abstract class BlockEntityMixin implements DynamicLightSource
             }
 
             // Schedules the rebuild of chunks.
-            this.lambdynlights_scheduleTrackedChunksRebuild(renderer);
+            this.ryoamicLights$scheduleTrackedChunksRebuild(renderer);
             return true;
         }
         return false;
     }
 
     @Override
-    public void lambdynlights_scheduleTrackedChunksRebuild(@NotNull WorldRenderer renderer)
+    public void ryoamicLights$scheduleTrackedChunksRebuild(@NotNull WorldRenderer renderer)
     {
         if (this.world == MinecraftClient.getInstance().world)
         for (long pos : this.trackedLitChunkPos) {