+++ src/main/java/me/lambdaurora/lambdynlights/mixin/lightsource/LivingEntityMixin.java
@@ -31,7 +31,7 @@ public abstract class LivingEntityMixin extends Entity implements DynamicLightSo
     }
 
     @Override
-    public void dynamicLightTick()
+    public void ryoamicLights$dynamicLightTick()
     {
         if (this.isOnFire() || this.isGlowing()) {
             this.lambdynlights_luminance = 15;
@@ -56,7 +56,7 @@ public abstract class LivingEntityMixin extends Entity implements DynamicLightSo
     }
 
     @Override
-    public int getLuminance()
+    public int ryoamicLights$getLuminance()
     {
         return this.lambdynlights_luminance;
     }