+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Changelog
-- fix forge version in modinfo ([#1](https://github.com/ThinkingStudios/LambDynamicLights-Forge/issues/1))
+- change modid to `ryoamiclights`
+- fix compatible with `include LDL` mods (maybe? idk).
+- use `renderWorldLastEvent`.